chore: updated component to encapsulate 2 screens
